converted pages content to modules
