+++ next-client/components/ContentWrapper.js
@@ -1,7 +1,10 @@
-const ContentWrapper = ({ children }) => {
+const ContentWrapper = ({ 
+  children,
+  size = "2xl"
+}) => {
   return (
     <main className="w-screen px-8 pt-16">
-      <div className="max-w-2xl mx-auto pb-20">
+      <div className={`max-w-${size} mx-auto pb-20`}>
         {children}
       </div>
     </main>
