+++ next-client/modules/entry/NewEntryPage.js
@@ -1,0 +1,11 @@
+import EntryBuilder from "../../components/EntryBuilder"
+
+const NewEntryPage = () => {
+  return (
+    <>
+      <EntryBuilder />
+    </>
+  )
+}
+
+export default NewEntryPage
